V5.3.39

+++ app/Jobs/Util/UnlinkFile.php
@@ -39,6 +39,10 @@ class UnlinkFile implements ShouldQueue
      */
     public function handle()
     {
+        /* Do not delete files if we are on the sync queue*/
+        if(config('queue.default') == 'sync')
+            return;
+
         Storage::disk($this->disk)->delete($this->file_path);
     }
 }
